Created and authToken module so auth can be used easily in many controllers by requiring the function
