+++ authToken.js
@@ -1,0 +1,13 @@
+function authToken(req, res, next) {
+	const authHeader = req.headers['authorization'];
+	const token = authHeader && authHeader.split(' ')[1];
+	if (token == null) return res.send('no token');
+
+	jwt.verify(token, process.env.ACCESS_TOKEN_SECRET, (err, user) => {
+		if (err) return res.send('no access but has token');
+		req.user = user;
+		next();
+	});
+}
+
+export default authToken
