+++ controllers/posts.js
@@ -1,4 +1,6 @@
+require('dotenv').config();
 const express = require('express');
+const jwt = require('jsonwebtoken');
 
 const router = express.Router();
